requirements.txt
Django>=5.2,<5.3
Pillow>=11.0
django-markdownx>=4.0
html5lib-truncation>=0.1
gunicorn>=23.0
celery[redis]>=5.4
python-nmap>=0.7
